+++ src/io/hdf5_aux.h
@@ -49,10 +49,10 @@ H5L_type_t getLinkType(H5::H5Object &obj);
 /* convenience functions */
 H5::DataSet openOrCreateDataSet(H5::CommonFG& cfg, const char *name, H5::DataType type, H5::DataSpace space);
 H5::DataSet openOrCreateDataSet(H5::CommonFG& cfg, std::string name, H5::DataType type, H5::DataSpace space);
-H5::Group openOrCreateGroup(H5::CommonFG& cfg, const char *name, int size = 0);
-H5::Group openOrCreateGroup(H5::CommonFG& cfg, std::string name, int size = 0);
-H5::Group clearOrCreateGroup(H5::CommonFG& cfg, const char *name, int size = 0);
-H5::Group clearOrCreateGroup(H5::CommonFG& cfg, std::string name, int size = 0);
+H5::Group openOrCreateGroup(H5::CommonFG& cfg, const char *name, size_t size = 0);
+H5::Group openOrCreateGroup(H5::CommonFG& cfg, std::string name, size_t size = 0);
+H5::Group clearOrCreateGroup(H5::CommonFG& cfg, const char *name, size_t size = 0);
+H5::Group clearOrCreateGroup(H5::CommonFG& cfg, std::string name, size_t size = 0);
 
 H5::Group inline openGroup(hid_t gid, const char *name) {
     hid_t newGroup = H5Gopen(gid, name, H5P_DEFAULT);
